Add recoil compensation, recovery cancellation, tick optimization, and fix incomplete crosshair recovery bug
