+++ Source/CrystalRecoilEditor/Public/AssetTools/CRAssetTypeActions_RecoilPattern.h
