+++ Source/CrystalRecoilEditor/Public/AssetTools/CRRecoilPatternFactory.h
