+++ Source/CrystalRecoilEditor/Public/Widget/CRRecoilUnitGraphEditor.h
@@ -3,10 +3,10 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "CRRecoilUnitGraph.h"
-#include "CRRecoilUnitGraphWidgetDragOperations.h"
-#include "Misc/StringOutputDevice.h"
 #include "Widgets/SCompoundWidget.h"
+#include "Misc/StringOutputDevice.h"
+#include "Data/CRRecoilUnitGraph.h"
+#include "Editor/CRRecoilUnitGraphWidgetDragOperations.h"
 #include "CRRecoilUnitGraphEditor.generated.h"
 
 class FCRRecoilUnitSelection;
@@ -33,6 +33,7 @@ USTRUCT()
 struct FCRRecoilUnitClipboardData
 {
 	GENERATED_BODY()
+
 	bool ImportFromString(const FString& ImportString)
 	{
 		FStringOutputDevice errors;
@@ -60,64 +61,92 @@ public:
 		Axis_Y
 	};
 
-	/** Constructs this widget with InArgs */
 	void Construct(const FArguments& InArgs);
-	void SetObject(UCRRecoilUnitGraph* InRecoilUnitGraph);
-	void RegisterCommands() const;
+
+	void SetRecoilUnitGraph(UCRRecoilUnitGraph* InRecoilUnitGraph);
 
 	// SWidget interface
 	virtual int32 OnPaint(const FPaintArgs& Args, const FGeometry& AllottedGeometry, const FSlateRect& MyCullingRect, FSlateWindowElementList& OutDrawElements, int32 LayerId, const FWidgetStyle& InWidgetStyle, bool bParentEnabled) const override;
+
 	virtual bool SupportsKeyboardFocus() const override { return true; }
+
 	virtual FReply OnMouseButtonDown(const FGeometry& MyGeometry, const FPointerEvent& MouseEvent) override;
+
 	virtual FReply OnMouseButtonUp(const FGeometry& MyGeometry, const FPointerEvent& MouseEvent) override;
+
 	virtual FReply OnMouseMove(const FGeometry& MyGeometry, const FPointerEvent& MouseEvent) override;
+
 	virtual FReply OnMouseWheel(const FGeometry& MyGeometry, const FPointerEvent& MouseEvent) override;
+
 	virtual void OnMouseLeave(const FPointerEvent& MouseEvent) override;
+
 	virtual FReply OnKeyUp(const FGeometry& MyGeometry, const FKeyEvent& InKeyEvent) override;
 	// End of SWidget interface
 
 	void DrawOrigin(FSlateWindowElementList& OutDrawElements, const FGeometry& AllottedGeometry, const int32 BaseLayerID) const;
+
 	void DrawRecoilUnits(FSlateWindowElementList& OutDrawElements, const FGeometry& AllottedGeometry, int32 BaseLayerID) const;
+
 	void DrawGridAxisNumbers(FSlateWindowElementList& OutDrawElements, const FGeometry& AllottedGeometry, const int32 BaseLayerID) const;
+
 	void DrawSingleGridAxisNumber(const int32 LineIndex, const float OffsetFromGraphOriginToWidgetCenter, const EAxis Axis, const FSlateFontInfo& NumberFontInfo, const TSharedRef<FSlateFontMeasure>& FontMeasure, const int32 GridAxisStep, const FGeometry& AllottedGeometry, FSlateWindowElementList& OutDrawElements, const int32 BaseLayerID) const;
+
 	void DrawSelectionBox(FSlateWindowElementList& OutDrawElements, const FGeometry& AllottedGeometry, const int32 BaseLayerID) const;
 
 	void AddUnitUnderCursor() const;
+
 	void AddUnit(const FVector2f& RecoilLocation) const;
+
 	void ZoomToFitAllUnits() const;
+
 	void CopySelectedUnits() const;
+
 	void PasteUnits() const;
+
 	void DeleteUnits() const;
+
 	void StartUnitScaling();
+
 	void StopUnitScaling();
 
 protected:
 	UCRRecoilUnitGraph* GetUnitGraph() const;
+
 	UCRRecoilPattern* GetRecoilPattern() const;
+
 	FCRRecoilUnitSelection& GetRecoilUnitSelection() const;
 
 	FVector2f RecoilCoordsToGraphCoords(FVector2f RecoilCoords) const;
+
 	FVector2f GraphCoordsToRecoilCoords(FVector2f GraphCoords) const;
 
 	int32 FindUnitByScreenLocation(const FVector2f& ScreenLocation) const;
 
 	void SelectUnitsInPanelCoordsRect(const FSlateRect& SelectionRect) const;
 
+	void TryAutoRearrangeUnits() const;
+
 	int32 ScaleFromRecoilCoordsToGraphCoords = 16;
 
-	bool bNeedZoomToFit = false;
+	TWeakObjectPtr<UCRRecoilUnitGraph> RecoilUnitGraph;
 
-	UCRRecoilUnitGraph* RecoilUnitGraph = nullptr;
 	FCRRecoilPatternEditor* RecoilPatternEditor = nullptr;
 
 	TSharedPtr<SCRRecoilUnitGraphBackgroundWidget> BackgroundWidget;
+
 	TSharedPtr<SVerticalBox> ShortcutsContainer;
 
 	TOptional<FCRUnitGraphViewDelayedDrag> ViewDrag;
+
 	TOptional<FCRUnitGraphSelectionDelayedDrag> SelectionDrag;
+
 	TOptional<FCRUnitGraphMoveUnitsDelayedDrag> MoveUnitsDrag;
+
 	TOptional<FCRUnitGraphScaleUnitsDelayedDrag> ScaleUnitsDrag;
 
 	int32 LastLeftMouseDownFoundUnitID = INDEX_NONE;
+
 	FVector2f CurrentMousePanelPosition;
+
+	mutable bool bNeedZoomToFit = false;
 };
